requirements.txt
Django==3.2.14
Pillow==9.2.0
django-admin-sortable2==2.1
django-tinymce==3.4.0
requests==2.28.1
